@@ -46,23 +46,29 @@
     "url-loader": "^4.1.1",
     "vue-hot-reload-api": "^2.3.4",
     "vue-html-loader": "^1.2.4",
-    "vue-loader": "^15.9.6",
+    "vue-loader": "^16.1.2",
     "vue-style-loader": "^4.1.2",
     "vue-template-compiler": "^2.6.12",
-    "webpack": "^5.13.0",
+    "webpack": "^5.14.0",
     "webpack-bundle-analyzer": "^4.3.0",
     "webpack-cli": "^4.3.1",
-    "webpack-dev-server": "^3.11.1",
+    "webpack-dev-server": "^3.11.2",
     "webpack-merge": "^5.7.3"
   },
   "dependencies": {
     "@babel/core": "^7.12.10",
+    "@vue/compiler-sfc": "^3.0.5",
     "compression-webpack-plugin": "^7.1.2",
+    "imagemin-gifsicle": "^7.0.0",
+    "imagemin-mozjpeg": "^9.0.0",
+    "imagemin-pngquant": "^9.0.1",
+    "imagemin-webpack-plugin": "^2.4.2",
+    "img-loader": "^3.0.2",
     "mini-css-extract-plugin": "^1.3.3",
     "raw-loader": "^1.0.0",
     "terser-webpack-plugin": "^5.1.1",
-    "vue": "^2.6.12",
-    "vue-router": "^3.4.9"
+    "vue": "^3.0.5",
+    "vue-router": "^4.0.3"
   },
   "browserslist": [
     "defaults",
