@@ -1,10 +1,9 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import Index from '@/components/index/index'
-import './index.scss'
-
 import router from '@/router/index'
+import '@/css/common.scss'
 
-new Vue({
-    router,
-    render: h => h(Index)
-}).$mount('#index')
+const app = createApp({})
+app.component('index',Index)
+app.use(router)
+app.mount('#index')
