@@ -1,12 +1,7 @@
-import Vue from 'vue'
+import { createApp } from 'vue'
 import Index from '@/components/page2/index'
 import './page2.scss'
 import '../../assets/font/iconfont.css'
 
-new Vue({
-    el: '#page2',
-    render: h => h(Index),
-    // components: {
-    //     'index2': Index
-    // }
-})
+const app = createApp(Index)
+app.mount('#page2')
