@@ -6,15 +6,19 @@
 </template>
 
 <script>
+import router from "@/router";
     export default {
-        methods: {
-            showSection3() {
-                this.$emit('showSection3')
-            }
+      setup() {
+        const showSection3 = () => {
+          router.push('/ccc')
         }
+        return {
+          showSection3
+        }
+      }
     }
 </script>
 
-<style scoped>
-
+<style lang="scss">
+@import "src/css/common.scss";
 </style>
